services/document-processor/requirements.txt
# Core dependencies
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Azure services
azure-ai-formrecognizer==3.3.2
azure-core==1.29.5
openai==1.54.0
httpx==0.27.0

# Storage
minio==7.2.3
psycopg2-binary==2.9.9

# Message queue
pika==1.3.2

# Language detection
langdetect==1.0.9

# Text processing
tiktoken==0.5.2
markdown==3.5.1

# Utilities
python-json-logger==2.0.7
